Refactor app component
- Moved Group class to it's own file
- Editor components into editor folder
