+++ src/components/UploadFileDialog.tsx
@@ -22,8 +22,8 @@
 import * as React from "react";
 import { Service } from "../service";
 import * as ReactModal from "react-modal";
-import { Button } from "./Button";
-import { GoGear, GoFile, GoX, Icon, GoPencil, GoCheck } from "./Icons";
+import { Button } from "./shared/Button";
+import { GoGear, GoFile, GoX, Icon, GoPencil, GoCheck } from "./shared/Icons";
 import {File, FileType, Directory, extensionForFileType, nameForFileType, filetypeForExtension} from "../model";
 import { KeyboardEvent, ChangeEvent, ChangeEventHandler } from "react";
 import { ListBox, ListItem, FileUploadInput } from "./Widgets";
